+++ src/Entity/Embeddable/Detail.php
@@ -4,10 +4,18 @@ namespace App\Entity\Embeddable;
 
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Gedmo\Timestampable\Traits\TimestampableEntity;
 
 #[ORM\Embeddable]
 class Detail
 {
+    /**
+     * Hook timestampable behavior
+     * updates createdAt, updatedAt fields
+     */
+    use TimestampableEntity;
+
+
     #[ORM\Column(length: 255)]
     private ?string $title = null;
 
@@ -22,13 +30,6 @@ class Detail
 
     #[ORM\Column(length: 255)]
     private ?string $shortDescription = null;
-
-    #[ORM\Column(type: Types::DATETIME_MUTABLE)]
-    private ?\DateTimeInterface $createdAt = null;
-
-    #[ORM\Column(type: Types::DATETIME_MUTABLE, nullable: true)]
-    private ?\DateTimeInterface $updatedAt = null;
-
 
     public function getTitle(): ?string
     {
@@ -86,30 +87,6 @@ class Detail
     public function setShortDescription(string $shortDescription): self
     {
         $this->shortDescription = $shortDescription;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?\DateTimeInterface
-    {
-        return $this->createdAt;
-    }
-
-    public function setCreatedAt(\DateTimeInterface $createdAt): self
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeInterface
-    {
-        return $this->updatedAt;
-    }
-
-    public function setUpdatedAt(?\DateTimeInterface $updatedAt): self
-    {
-        $this->updatedAt = $updatedAt;
 
         return $this;
     }