+++ src/Form/DetailType.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Embeddable\Detail;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\DateType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class DetailType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('startAt', DateType::class, [
+                'required' => true
+            ])
+            ->add('endAt', DateType::class, [
+                'required' => false
+            ])
+            ->add('title', TextType::class, [
+                'required' => true
+            ])
+            ->add('shortDescription', TextType::class, [
+                'required' => true
+            ])
+            ->add('description', TextareaType::class, [
+                'required' => false
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Detail::class,
+        ]);
+    }
+}